+++ src/Header/AcceptEncoding.php
@@ -3,26 +3,43 @@ declare(strict_types = 1);
 
 namespace Innmind\Http\Header;
 
-use Innmind\Http\Header;
+use Innmind\Http\{
+    Header,
+    Header\Accept\Encoding,
+};
+use Innmind\Immutable\Sequence;
 
 /**
  * @psalm-immutable
  */
 final class AcceptEncoding implements Custom
 {
-    private Header $header;
+    /**
+     * @param Sequence<Encoding> $encodings
+     */
+    private function __construct(
+        private Sequence $encodings,
+    ) {
+    }
 
     /**
+     * @psalm-pure
      * @no-named-arguments
      */
-    public function __construct(AcceptEncodingValue ...$values)
+    public static function of(Encoding ...$encodings): self
     {
-        $this->header = new Header('Accept-Encoding', ...$values);
+        return new self(Sequence::of(...$encodings));
     }
 
     #[\Override]
     public function normalize(): Header
     {
-        return $this->header;
+        return new Header(
+            'Accept-Encoding',
+            ...$this
+                ->encodings
+                ->map(static fn($value) => new Value\Value($value->toString()))
+                ->toList(),
+        );
     }
 }